starting to build beverage fragment
This commit deos a bunch of refactoring so the main fragments are
clearly labeled as lists while the individual brewer and beverage
fragments can be named appropriately. This commit also begins work on
the beverage fragment. So far we are replacing the container framelayout
with a beverage fragment that populates the layout with beverage info.
That is all, no back funcitonality, no comments (though coming), no
interactivity.

There's work to be done for sure.