gradle upgrade, install realm, use activities
This commit creates and uses two different activities to present the
brewer and beverage pages. This change in anticipation of supporting
landscape mode on larger displays. It also changes the manifest so the
main activity is no reloaded on device rotation. Though this will have
to be handled in the future to present the dual pane layout to come.