introduce network change broadcast receiver
This commit introduces a network state change broadcast receiver to the
beverage list fragment. This receiver is used to capture changes in the
network. This receiver waits for changes in netowrk connectivity in the
case where data loading was unsuccessful for whatever reason, as soon
as connectivity is established, the data load is started. This keeps the
user from having to press the retry button, though the button will stay
in the UI to support uknown edge cases.

Further work needed is to put this receiver into the other two
fragments. However, since the fragments are so similar, we may be able
to consolidate the functionality into a base class.