add account sign up, refactor
This commit adds in an account sign up activity at the beginning of app
launch. This account is saved using Realm and will be later accessed by
the leaderboard and beverage fragments to determine a user's favorites
and comments on a given beverage. This commit also cleans up a few
classes and adds in Stetho for realm database viewing through Chrome.