add comment update
This commit adds the ability for for the beverage fragment to update the
comment content when the back button is pressed. As a note, we may want
to do the samething in the onPause or onStop  method in the case where
a user exits the app and the uses other apps, pushing ours out of memory.
Note that the comment update is done using a service that will live on
until the task is completed, whereas the lifetime of the fragment and
its activity has no gurantee.

This commit also configures realm in a more robust way, allowing the
realm file to be delete and recreated when the schema changes as it did
when we made Comment extend RealmObject.