add favorite updater to beverage fragment
This commit introduces, similar to comments, a service to save the state
of the favorite button when the user backs out of the beverage page.
This commit also introduces a Constants class that holds static values
including a method that builds Realm instances.

Work to do from here. Persist comment and favorite data so we don't have
to query for it when opening the beverage page and only update the data
when the user has changed it before backing out of the page.