save commets and favorites in db
This commit removes the network calls when opening the beverage page in
favour of saving/loading that information in the db.

If passwords are implemented in the future (which they should,
otherwise the database favorites and comments are useless) I will need
to check for comments and favorites and if they don't exist try a download
from the server. This will only occur in the case where a user has logged
out or deleted the app and has logged in or reinstalled the app. This is
a bit of an edge case considering the size of the app but it is sort of
real world.