fix autoincrement bug
This commit adds in some code to find the largest id value of the saved
comments so it can create and save new comments without getting an error
about a comment that already exists due to its primary key. It's kind of
lame that Realm doesn't have support for auto incremented values. Oh
well.